Module Carton.Blob

The Blob is a tuple of temporary buffers used to store an object that has been decompressed or reconstructed using a patch and a source.

The purpose of a Blob is to prepare the allocation of what is needed to store the object and (re)use this Blob throughout the reconstruction of the object (whether it's a base or a patch) in order to obtain the final object.

type t

The type of blobs.

val make : size:Size.t -> t

val of_string : string -> t

val size : t -> Size.t

val source : t -> De.bigstring

val with_source : t -> source:De.bigstring -> t

val payload : t -> De.bigstring

val flip : t -> t