Module Dec.Fp

First-pass of a PACK file.

From a stream, it is possible to infer information needed then to apply a second analyse to extract all objects of a the given PACK file. From Git, this pass appears when the client clone/fetch and the program counts how many objects the PACK file has.

$ git clone ...
remote: Enumerating objects: 105, done.
remote: Counting objects: 100% (105/105), done.
remote: Compressing objects: 100% (81/81), done.
remote: Total 305 (delta 41), reused 75 (delta 23), pack-reused 200
Receiving objects: 100% (305/305), 104.46 KiB | 0 bytes/s, done. # first pass

Parameters

module Uid : sig ... end

Signature

type optint = Optint.t

type kind =

| Base of [ `A | `B | `C | `D ]

| Ofs of {

sub : int;

source : weight;

target : weight;

}

| Ref of {

ptr : Uid.t;

source : weight;

target : weight;

}
(*
Type of PACK objects.
*)

type entry = {

offset : int64;
(*
Absolute offset into the given PACK file.
*)

kind : kind;
(*
Kind of the object.
*)

size : weight;
(*
Length of the inflated object.
*)

consumed : int;
(*
Length of the deflated object (as it is into the PACK file).
*)

crc : optint;
(*
Check-sum of the entry (header plus the deflated object).
*)

}

Type of a PACK entry.

val check_header : 's Carton__.Sigs.scheduler -> ('fd, 's) read -> 'fd -> (int * string * int, 's) Carton__.Sigs.io

type decoder

The type for decoders.

type src = [

| `Channel of Stdlib.in_channel

| `String of string

| `Manual

]

The type for input sources. With a `Manual source the client must provide input with src.

type decode = [

| `Await of decoder

| `Peek of decoder

| `Entry of entry * decoder

| `End of Uid.t

| `Malformed of string

]

type header =

| Consumed of Bigstringaf.t

| None

val decoder : o:Bigstringaf.t -> allocate:(int -> De.window) -> src -> decoder

val decode : decoder -> decode

val number : decoder -> int

val version : decoder -> int

val count : decoder -> int

val ctx : decoder -> Uid.ctx

val src_rem : decoder -> int

val src : decoder -> Bigstringaf.t -> int -> int -> decoder