Module Carton.Kind

A PACK file contains several types of object. According to Git, it contains commits (`A), trees (`B), blobs (`C) and tags (`D). Carton is far enough removed from Git to abstract itself from the actual type of these objects.

type t = [

| `A

| `B

| `C

| `D

]

The type of kinds of objects.

val pp : Stdlib.Format.formatter -> t -> unit

val compare : t -> t -> int

val equal : t -> t -> bool

val to_int : t -> int