Module Carton.Size

The size is a non-negative number which corresponds to the size of a Blob in memory in bytes.

Note: We're not talking about the size of an object in the PACK file, but the size of the buffers needed to extract an object. Indeed, depending on the object and the way it is serialized in the PACK file, it may be necessary for the size to be larger than the object itself, particularly to store the source needed to build a patched object.

type t = private int

val of_int_exn : int -> t

val to_int : t -> int

val zero : t

val compare : t -> t -> int

val equal : t -> t -> bool

val max : t -> t -> t