Module Carton.Uid

An object can be identified by a unique identifier that needs to be calculated by an algorithm such as a hash algorithm. These identifiers can be used to refer to a possible source when we have a First_pass.kind.Ref entry.

This means that all the Uids used must have strictly the same size, which corresponds in particular to ref_length — the value required to decode a PACK file.

type t = private string

val unsafe_of_string : string -> t

val pp : Stdlib.Format.formatter -> t -> unit

val compare : t -> t -> int

val equal : t -> t -> bool